fix edge-functions
